requirements.txt
boto3
logging
python-logmatic
